# Implement and optimize NormalQuantile