This PR moves the computation of the standard normal quantile from gonum/stat/distuv to gonum/mathext. This also optimizes the implementation that had been in distuv. While performing this modification, I also compared with the implementation in cephes, and applied the same optimizations. These optimizations may as well be kept.

The accuracy of this implementation and the implementation in cephes are similar (at least compared with the tests here).

At the start, the benchmark of the Cephes implementation was ~51000ns/op, reduced to ~41000ns/op after optimizations. The distuv code started at ~41000ns/op, and these optimizations reduce to 24600ns/op, about a 40% increase.

A piece of code I have been working on spends about 20% of the time just computing the normal quantile, and these optimizations lead to about a 10% improvement on this real program