@kortschak 2017-02-19
Would it be possible to make this two commits: 1) Move code from stat and 2) apply optimisations.